fix: incorrect zod def

+++ src/zod/query.ts
@@ -9,7 +9,7 @@ export const queryParams = z
     size: z.coerce.number().int().min(24).max(512),
     stroke: hexCode,
     background: hexCode,
-    backgroundAlpha: z.coerce.number().int().min(0).max(1),
+    backgroundAlpha: z.coerce.number().min(0).max(1),
     strokeWidth: z.preprocess(
       (arg) => parseFloat(z.string().parse(arg)),
       z.number().min(0.1).max(4)
